// lnd/rpc.d.ts
declare type SubtypeConstructor<Constructor extends new (...args: any) => any, Subtype> = {
    new (...args: ConstructorParameters<Constructor>): Subtype;
};
export interface ProtoGrpcType {
    lnrpc: {
// ...
        Lightning: SubtypeConstructor<typeof grpc.Client, _lnrpc_LightningClient> & {
            service: _lnrpc_LightningDefinition;
        };
// ...
    };
}