// lnd/walletunlocker.d.ts
declare type SubtypeConstructor<Constructor extends new (...args: any) => any, Subtype> = {
    new (...args: ConstructorParameters<Constructor>): Subtype;
};
export interface ProtoGrpcType {
    lnrpc: {
// ...
        Lightning: SubtypeConstructor<typeof grpc.Client, _lnrpc_LightningClient> & {
            service: _lnrpc_LightningDefinition;
        };
// ...
        WalletUnlocker: SubtypeConstructor<typeof grpc.Client, _lnrpc_WalletUnlockerClient> & {
            service: _lnrpc_WalletUnlockerDefinition;
        };
    };
}